Added cast to get rid of warning message. Original: 6f077c24b73b71cdc9fc8f49f65b84b675f35c13

+++ lal/packages/support/test/ReadFTSeriesTest.c
@@ -167,7 +167,7 @@ int main( void )
   UINT4   j;
 
   fprintf(stderr,"Testing value of LALUnitTextSize ... ");
-  if ( LALSupportUnitTextSize != LALUnitTextSize ) 
+  if ( (int)LALSupportUnitTextSize != (int)LALUnitTextSize ) 
   {
     fprintf(stderr,"UnitTextSize mismatch: [ReadFTSeriesTest:%s]\n",
 	    READFTSERIESTESTC_MSGEFLS);
